+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 ### Changed
+- Revert "services" to "modules" change.
+- Change "service" role into "module"
+- Add type field to services, which can be "module" or  "plugin"
+- Allow plugin users to create services with type "plugin"
+- Allow service-registerer users to create users with role "plugin"
 ### Fixed
 ### Removed